Fix: Remove AdmPwd.PS dependency from manifest
There is a module on the PSGallery shadowing the AdmPwd.PS module.
Since it is not the offical Microsoft module I can't be sure if there
isn't a bug/vulnerability in the code.

This change depends on AdmPwd.PS beeing installed in
`$env:windir\system32\WindowsPowerShell\v1.0\Modules\AdmPwd.PS`
which is the default directory in which the msi installer will put the
module.